Fix various bugs found with unit tests
ISeek::SetLoop was returning undefined status.
Khronos Bug 7614 - sizeof(long) != 32 bits on x86_64.
Fix bug in IID hashing due to the above.
Fix bug in GenericMediaPlayer::onLoop; was using mPlayer without checking for NULL.
Fix wrong API level in tests/sandbox/object.c.
Decrease log level from I to V in MediaPlayerNotificationClient::notify
    and GenericPlayer::~GenericPlayer.
Fix memory leak in tests/sandbox/engine.c.
Comment out overly verbose log in tests/sandbox/streamSource/slesTestPlayStream.cpp.

+++ wilhelm/include/OMXAL/OpenMAXAL_Platform.h
@@ -37,8 +37,8 @@ typedef unsigned char               xa_uint8_t;
 typedef signed char                 xa_int8_t;
 typedef unsigned short              xa_uint16_t;
 typedef signed short                xa_int16_t;
-typedef unsigned long               xa_uint32_t;
-typedef signed long                 xa_int32_t;
+typedef unsigned int /*long*/       xa_uint32_t;
+typedef signed int /*long*/         xa_int32_t;
 typedef long long                   xa_int64_t;
 typedef unsigned long long          xa_uint64_t;
